Store babel languages in the database
The current method of reading categories and parsing languages out of
them is a terrible hack, and a TODO comment suggests to use a database,
which is a much better structured storage format.

This adds a `babel` database table, which if $wgBabelUseDatabase is
true, will be used to store the languages a user knows, and at what
level.

Once the database is populated and created everywhere, we should remove
the fallback category code.

Change-Id: I520358ac778cfddb1cf768c798f322dce61922fd