Allow reading Babel information from a central API
If configured, when there is no local information in the Babel database,
it can fallback to a central API, and use that information instead.

This requires CentralIdLookup to be properly configured as well as the
remote wiki having a current version of the Babel extension.

Note that this will affect API responses on local wikis, as they
will display information from the central API.

Bug: T95877
Change-Id: Icdd56a8169b0a1bfeff82a13451e879a33a0ada4