Populate babel database table if it exists
The $wgBabelUseDatabase option is a temporary migration setting that
lets the new code exist alongside the old category-based one. However,
since it controls both reading from and writing to the database, you
can't populate the database ahead of time without also changing reads to
be from the (empty) database. So don't check the config setting for
writes, just populate the table as soon as it exists, and then allow
manually switching reads once it has been sufficiently populated.

Change-Id: Ia46ae6596cb9256145dbcc0b1371a1c7451f6aea