Invalidate local caches on central wiki edits
Currently the language cache is split per-wiki since we don't keep track
of where the information is coming from. If the local wiki is inheriting
from the central wiki, we need to invalidate the local cache whenever
the central wiki languages change.

We do this by using a global check key that uses the central user Id,
which is touched whenever the central wiki is edited.

Change-Id: I463a50494eb97b3c6040ae1563f326705eb06924