Differentiate MediaWiki-internal codes from BCP 47 codes
Keep BCP 47 and MediaWiki-internal codes separate in the code base, but
ensure that BCP 47 codes can be mapped to MediaWiki-internal codes if
they happen to be used.

Introduce a new "category code" format for backwards compatibility
with existing categories and/or database entries.  A "category code"
is a lowercased MediaWiki-internal code ("simple") *unless* the
MediaWiki-internal code contains a hyphen, in which case it is
formatted as in BCP 47 ("zh-Hant").  Pass around MediaWiki codes
everywhere for consistency, but make it clear where we are
converting to BCP 47 or "category code" format.

This patch should keep Babel from breaking when
Iebbc604af21d7f2af9c1f1ab2574cb5f309bf6ed lands in core, which
changes LanguageCode::bcp47("simple") from "simple" to "en-simple",
among other correctness fixes.

Bug: T199941
Change-Id: I27a5b8e45b34c6b57c1b612b11548001c88cd483