Improve tests to cover more language code corner cases
These tests would catch the regression found in T199941 by ensuring
that we test the behavior of nonstandard mediawiki language codes (not
BCP 47 compliant) like 'simple', as well as suffixed language codes
like 'zh-Hant'.  We also tweak the case in the test setup's
`{{#babel}}` invocation to ensure that we are properly normalizing
whatever crazy case the user is giving us.

Bug: T199941
Change-Id: I1c387aea3845565d88110eeb21b24201e2946b9d