tests: Use empty WANObjectCache instead of mock
It's not reliant on any global state and can be instantiated regularly
instead. Incedentally, this also means we can use variadic signatures
in WANObjectCache.php in core. This is currently failing because of
an obscure bug in HHVM that affects PHPUnit 4, which in turn makes
Phan unhappy.

The variadic feature itself works fine even in older HHVM on PHP5
mode and everything else we support. It's just that you can't call
createMock() on a class that uses it with PHPUnit 4, on HHVM.

Bug: T228563
Change-Id: Ifa5b96735376f2fbe3680799f960616ba8d357ff