Use a logically unique first segment for WANObjectCache keys
The first segment should be a single kind of key, this enables
better debugging via tools that monitor memcached network
traffic, as well as WANObjectCache's own Statsd statistics.

Avoid using the first merely as a high-level namespace, and avoid
using the same prefix for both a wiki-local key and a global key.

Change-Id: I8ecd7ed2f2ab989b116a673213d7a50fd2f17c8c