Remove static $user cache in BabelAutoCreate
Static caches can very easily lead to unpredictable test issues. In this
case, testUser() creates the user and leaves it in the cache. Then, when
testCreate() runs, if AbuseFilter is installed, it will create its own
system user. Assuming that user tables are cleared between each test
(which will be the case after T342301 is resolved), the AbuseFilter
user's ID will be 1, that was previously assigned to the Babel user.
This leads to two different users having the same ID, which fails when
the second user is created.

To avoid any chance of other unrelated tests failing, just delete the
static cache. It will be possible to add it again and make it non-static
when (and if!) BabelAutoCreate is updated to use DI.

Change-Id: I53e133d887212072a514850e70a7ee95b64458db