Make the Babel class instantiable
Storing $title in a static property looks scary – what if two parses
happen at the same time? For followup commits, storing the parser is
convenient, but even more scary. So make the Babel class instantiable,
make $title an instance property, and also store the parser in an
instance property. Methods that access neither the title nor the parser
remain static, as there are public static functions anyway (which can’t
be instance methods, as they don’t provide a parser).

Change-Id: I3d88fb5cae5384f8a2b6cfe3dd4028b32dc0a28f