Use Parser::getPage() directly, don’t convert to Title
Most usage is fine with the restricted API PageReference provides. There
is a single place (and even that one doesn’t run with the default
configuration) where Title is much more convenient; convert
PageReference to Title there instead of at the top level.

This also means that the coding convention [1] of not using Title in
type declarations is now followed.

[1] https://www.mediawiki.org/wiki/Manual:Coding_conventions/PHP#Don't_add_type_declarations_for_%22big%22_legacy_classes

Bug: T379934
Change-Id: I57cd539f14a0bd2869f45964f94ac66ba517770c